# packages/shared/openapi/openapi-map.json
{
  "components": {
    "schemas": {
      "ActivityStatsOut": {
        "additionalProperties": false,
        "description": "Activity-level aggregates (Engine 3).",
        "properties": {
          "avg_speed_kmh": {
            "minimum": 0.0,
            "title": "Avg Speed Kmh",
            "type": "number"
          },
          "moving_time_s": {
            "description": "Wall-clock moving duration in seconds",
            "minimum": 0.0,
            "title": "Moving Time S",
            "type": "number"
          },
          "top_speed_kmh": {
            "minimum": 0.0,
            "title": "Top Speed Kmh",
            "type": "number"
          },
          "total_distance_km": {
            "minimum": 0.0,
            "title": "Total Distance Km",
            "type": "number"
          },
          "total_vertical_drop_m": {
            "title": "Total Vertical Drop M",
            "type": "number"
          },
          "trail_count": {
            "description": "Distinct named trails in the activity",
            "minimum": 0.0,
            "title": "Trail Count",
            "type": "integer"
          }
        },
        "required": [
          "total_distance_km",
          "total_vertical_drop_m",
          "top_speed_kmh",
          "avg_speed_kmh",
          "moving_time_s",
          "trail_count"
        ],
        "title": "ActivityStatsOut",
        "type": "object"
      },
      "ElevationCorrectionRequest": {
        "additionalProperties": false,
        "description": "Correct or fill elevations for a sequence of track points.",
        "properties": {
          "points": {
            "items": {
              "$ref": "#/components/schemas/TrackPointIn"
            },
            "maxItems": 512,
            "minItems": 1,
            "title": "Points",
            "type": "array"
          }
        },
        "required": [
          "points"
        ],
        "title": "ElevationCorrectionRequest",
        "type": "object"
      },
      "ElevationCorrectionResponse": {
        "additionalProperties": false,
        "properties": {
          "points": {
            "items": {
              "$ref": "#/components/schemas/TrackPointOut"
            },
            "title": "Points",
            "type": "array"
          }
        },
        "required": [
          "points"
        ],
        "title": "ElevationCorrectionResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MapActivityCreateRequest": {
        "additionalProperties": false,
        "description": "Persist a full pipeline activity: header + ordered points + segment slices.",
        "properties": {
          "processed_track": {
            "$ref": "#/components/schemas/ProcessedTrackIn"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/SegmentOut-Input"
            },
            "title": "Segments",
            "type": "array"
          },
          "stats": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "$ref": "#/components/schemas/ActivityStatsOut"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stats"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "processed_track"
        ],
        "title": "MapActivityCreateRequest",
        "type": "object"
      },
      "MapActivityDetailResponse": {
        "additionalProperties": false,
        "description": "Activity row plus reconstructed processed track and segments (indices are inclusive).",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "processed_track": {
            "$ref": "#/components/schemas/ProcessedTrackOut"
          },
          "recorded_at": {
            "format": "date-time",
            "title": "Recorded At",
            "type": "string"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/SegmentOut-Output"
            },
            "title": "Segments",
            "type": "array"
          },
          "stats": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stats"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Url"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user_id",
          "recorded_at",
          "processed_track",
          "segments"
        ],
        "title": "MapActivityDetailResponse",
        "type": "object"
      },
      "MapActivityListItem": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "recorded_at": {
            "format": "date-time",
            "title": "Recorded At",
            "type": "string"
          },
          "stats": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stats"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user_id",
          "recorded_at"
        ],
        "title": "MapActivityListItem",
        "type": "object"
      },
      "MapActivityListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MapActivityListItem"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "MapActivityListResponse",
        "type": "object"
      },
      "PointSegmentType": {
        "enum": [
          "lift",
          "run",
          "transition"
        ],
        "title": "PointSegmentType",
        "type": "string"
      },
      "ProcessedTrackIn": {
        "additionalProperties": false,
        "description": "Engine 1 normalized track (input).",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/TrackPointIn"
            },
            "minItems": 1,
            "title": "Points",
            "type": "array"
          },
          "recorded_at": {
            "format": "date-time",
            "title": "Recorded At",
            "type": "string"
          },
          "source_type": {
            "$ref": "#/components/schemas/SourceType"
          }
        },
        "required": [
          "id",
          "points",
          "recorded_at",
          "source_type"
        ],
        "title": "ProcessedTrackIn",
        "type": "object"
      },
      "ProcessedTrackOut": {
        "additionalProperties": false,
        "description": "Engine 1 normalized track (output).",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/TrackPointOut"
            },
            "minItems": 1,
            "title": "Points",
            "type": "array"
          },
          "recorded_at": {
            "format": "date-time",
            "title": "Recorded At",
            "type": "string"
          },
          "source_type": {
            "$ref": "#/components/schemas/SourceType"
          }
        },
        "required": [
          "id",
          "points",
          "recorded_at",
          "source_type"
        ],
        "title": "ProcessedTrackOut",
        "type": "object"
      },
      "SegmentOut-Input": {
        "additionalProperties": false,
        "description": "Engine 2 classified slice.",
        "properties": {
          "difficulty": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Difficulty"
          },
          "end_index": {
            "minimum": 0.0,
            "title": "End Index",
            "type": "integer"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/TrackPointOut"
            },
            "title": "Points",
            "type": "array"
          },
          "start_index": {
            "minimum": 0.0,
            "title": "Start Index",
            "type": "integer"
          },
          "trail_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trail Name"
          },
          "type": {
            "$ref": "#/components/schemas/SegmentType"
          }
        },
        "required": [
          "type",
          "points",
          "start_index",
          "end_index"
        ],
        "title": "SegmentOut",
        "type": "object"
      },
      "SegmentOut-Output": {
        "additionalProperties": false,
        "description": "Engine 2 classified slice.",
        "properties": {
          "difficulty": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Difficulty"
          },
          "end_index": {
            "minimum": 0.0,
            "title": "End Index",
            "type": "integer"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/TrackPointOut"
            },
            "title": "Points",
            "type": "array"
          },
          "start_index": {
            "minimum": 0.0,
            "title": "Start Index",
            "type": "integer"
          },
          "trail_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trail Name"
          },
          "type": {
            "$ref": "#/components/schemas/SegmentType"
          }
        },
        "required": [
          "type",
          "points",
          "start_index",
          "end_index"
        ],
        "title": "SegmentOut",
        "type": "object"
      },
      "SegmentType": {
        "enum": [
          "descent",
          "lift",
          "flat",
          "pause"
        ],
        "title": "SegmentType",
        "type": "string"
      },
      "SourceType": {
        "enum": [
          "gpx",
          "fit",
          "live"
        ],
        "title": "SourceType",
        "type": "string"
      },
      "TrackPointIn": {
        "additionalProperties": false,
        "description": "Atomic GPS sample (input). `elevation_m` optional when requesting DEM correction.\n\nDEM Correction: process that uses a digital elevation model (DEM) to correct the elevation of a track point.",
        "properties": {
          "elevation_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elevation M"
          },
          "lat": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lon",
            "type": "number"
          },
          "segment_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PointSegmentType"
              },
              {
                "type": "null"
              }
            ]
          },
          "speed_kmh": {
            "minimum": 0.0,
            "title": "Speed Kmh",
            "type": "number"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "lat",
          "lon",
          "timestamp",
          "speed_kmh"
        ],
        "title": "TrackPointIn",
        "type": "object"
      },
      "TrackPointOut": {
        "additionalProperties": false,
        "description": "Atomic GPS sample (output); elevation from sensor or corrected.\nAfter DEM correction is applied:",
        "properties": {
          "elevation_m": {
            "title": "Elevation M",
            "type": "number"
          },
          "lat": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lon",
            "type": "number"
          },
          "segment_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PointSegmentType"
              },
              {
                "type": "null"
              }
            ]
          },
          "speed_kmh": {
            "minimum": 0.0,
            "title": "Speed Kmh",
            "type": "number"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "lat",
          "lon",
          "elevation_m",
          "timestamp",
          "speed_kmh"
        ],
        "title": "TrackPointOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "_ThumbnailRequest": {
        "properties": {
          "activity_id": {
            "title": "Activity Id",
            "type": "string"
          },
          "points": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Points",
            "type": "array"
          }
        },
        "required": [
          "activity_id",
          "points"
        ],
        "title": "_ThumbnailRequest",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "## Canonical API surface\n\nAll supported clients MUST use **`/api/v1/*`** paths documented here.",
    "title": "Map Backend API",
    "version": "1.0.0",
    "x-canonical-prefix": "/api/v1/map"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/map/activities": {
      "get": {
        "operationId": "list_activities_api_v1_map_activities_get",
        "parameters": [
          {
            "description": "Filter activities for this user",
            "in": "query",
            "name": "user_id",
            "required": true,
            "schema": {
              "description": "Filter activities for this user",
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MapActivityListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Activities",
        "tags": [
          "activities"
        ]
      },
      "post": {
        "operationId": "create_activity_api_v1_map_activities_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MapActivityCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MapActivityDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Activity",
        "tags": [
          "activities"
        ]
      }
    },
    "/api/v1/map/activities/thumbnail": {
      "post": {
        "description": "Generate and upload a route thumbnail from raw GPS points (live recording path).",
        "operationId": "generate_activity_thumbnail_api_v1_map_activities_thumbnail_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_ThumbnailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Generate Activity Thumbnail Api V1 Map Activities Thumbnail Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Activity Thumbnail",
        "tags": [
          "activities"
        ]
      }
    },
    "/api/v1/map/activities/{activity_id}": {
      "delete": {
        "description": "Delete activity header; ``track_points`` and ``segments`` cascade via FK.",
        "operationId": "delete_activity_api_v1_map_activities__activity_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "activity_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Activity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Activity",
        "tags": [
          "activities"
        ]
      },
      "get": {
        "operationId": "get_activity_api_v1_map_activities__activity_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "activity_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Activity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MapActivityDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Activity",
        "tags": [
          "activities"
        ]
      }
    },
    "/api/v1/map/elevation/correct": {
      "post": {
        "description": "Correct ``elevation_m`` using Copernicus GLO-30 DEM tiles (local cache + rasterio).\n\nResponses are cached by **quantized bounding box** and **per-point (lat, lon) signature**\nfor the lifetime of the process (LRU + TTL).",
        "operationId": "correct_elevation_dem_api_v1_map_elevation_correct_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ElevationCorrectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ElevationCorrectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Correct Elevation Dem",
        "tags": [
          "elevation-dem"
        ]
      }
    },
    "/api/v1/map/trails/resort": {
      "get": {
        "description": "GeoJSON ``FeatureCollection`` of ``map_trail.ski_runs`` lines intersecting the bbox.\n\nUsed by MapLibre to prefetch a resort vector layer. For segment enrichment,\ncall the theta trail server instead of map-backend.",
        "operationId": "resort_trails_geojson_api_v1_map_trails_resort_get",
        "parameters": [
          {
            "description": "Bounding box WGS84: min_lon,min_lat,max_lon,max_lat",
            "in": "query",
            "name": "bbox",
            "required": true,
            "schema": {
              "description": "Bounding box WGS84: min_lon,min_lat,max_lon,max_lat",
              "examples": [
                "8.0,47.0,9.0,48.0"
              ],
              "title": "Bbox",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Resort Trails Geojson Api V1 Map Trails Resort Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resort Trails Geojson",
        "tags": [
          "trails"
        ]
      }
    }
  }
}